avoided adding run instance into runs list on phone button. change it later. And refactored some file's name.
